Fixed a bug causing the boss to invoke its attacks twice.
The interrupt timer wasn't running properly and finished before it had the chance to check if it was interrupted or not. So in the case of it being hit during that cycle, it would reactivate after the head drops, then once again after the invoke timer completes and calls the attack cycles a second time.